requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2

# Celery & Redis
celery==5.4.0
redis==5.0.1
django-redis==5.4.0
django-filter==24.3
drf-yasg==1.21.8
gunicorn==23.0.0
whitenoise==6.7.0
